# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Documentation",
        "description": "Server endpoint for Casper project. Each of the resources below has multiple endpoints which allows end users and developers to access the Gunther API. This server allows you can use the api key  \"special-key\" to test the authorization filters",
        "termsOfService": "http://localhost",
        "contact": {
            "email": ""
        },
        "license": {
            "name": "",
            "url": ""
        },
        "version": "0.0.1"
    },
    "host": "localhost:9000",
    "basePath": "/v1/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/orders": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Get all orders",
                "description": "Returns all orders.",
                "operationId": "get.orders",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Retrieve all orders.",
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    },
                    "404": {
                        "description": "Error fetching order.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
        "/orders/{id}/valid": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Verifies if valid order ID",
                "description": "Checks if order is present in current ordered items in DB",
                "operationId": "validate.order.for.id",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of specific order that needs to be verified",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Verifies if valid order ID",
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    },
                    "404": {
                        "description": "Error fetching order.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
        "/orders/{id}": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Retrieve a specific existing order",
                "description": "Returns an order based on a specific order ID",
                "operationId": "get.order.for.id",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of specific order that needs to be retrieved",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retrieve order based on a specific order ID",
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    },
                    "404": {
                        "description": "Error fetching order.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Order"
                ],
                "summary": "Delete a specific existing order",
                "description": "Deletes order based on a specific order ID",
                "operationId": "delete.order.for.id",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of specific order that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "404": {
                        "description": "Error creating order",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
        "/user/{id}/orders": {
            "get": {
                "tags": [
                    "Order"
                ],
                "summary": "Retrieve a existings order for user",
                "description": "Returns all orders based on a specific user ID",
                "operationId": "get.orders.for.user.via.id",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of user for orders to be retrieved",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retrieve order based on a specific user ID",
                        "schema": {
                            "$ref": "#/definitions/Order"
                        }
                    },
                    "404": {
                        "description": "Error fetching order.",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
        "/orders/add": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Create a order",
                "description": "Creates a new order",
                "operationId": "create.order",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "userId",
                        "in": "formData",
                        "description": "Unique Id of user requesting an order",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "amount",
                        "in": "formData",
                        "description": "Total amount of order",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "transactionId",
                        "in": "formData",
                        "description": "Transaction Id assocaited with Stripe",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Displays new order created",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "404": {
                        "description": "Error creating order",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
        "/orders/{id}/add/{itemId}": {
            "post": {
                "tags": [
                    "Order"
                ],
                "summary": "Creates a order item",
                "description": "Creates a new order item",
                "operationId": "create.order.item",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of specific order that needs to be retrieved",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "itemId",
                        "in": "path",
                        "description": "ID of specific product to be added to the order",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "List of order",
                        "schema": {
                            "$ref": "#/definitions/OrderItem"
                        }
                    },
                    "404": {
                        "description": "Error creating order",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
        "/orders/{id}/delete": {
            "delete": {
                "tags": [
                    "Order"
                ],
                "summary": "Delete a specific existing order item",
                "description": "Deletes order item based on a specific ID",
                "operationId": "delete.order.item.for.id",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of specific order item that needs to be deleted",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "404": {
                        "description": "Error deleting order item",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        },
...
        "/products/add": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create a product",
                "description": "Creates a new product",
                "operationId": "create.product",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "Name of product to be created",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "description": "Product description",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "price",
                        "in": "formData",
                        "description": "Price of product to be created",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "imageurl",
                        "in": "formData",
                        "description": "Image url of product to be created",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "file",
                        "in": "formData",
                        "description": "Product image file to upload",
                        "required": false,
                        "type": "file"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of products",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    },
                    "404": {
                        "description": "Error creating product",
                        "schema": {
                            "$ref": "#/definitions/ErrorModel"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Order": {
            "required": [
                "user_id",
                "total_paid",
                "stripe_transaction_id"
            ],
            "properties": {
                "id": {
                    "description": "Unique order identifier",
                    "type": "string"
                },
                "user_id": {
                    "description": "User_id related to existing order",
                    "type": "string"
                },
                "total_paid": {
                    "description": "Total amount paid for order",
                    "type": "string"
                },
                "stripe_transaction_id": {
                    "description": "Transaction id related to existing order",
                    "type": "string"
                }
            },
            "type": "object",
            "xml": {
                "name": "Order"
            },
            "externalDocs": {
                "description": "find more info here",
                "url": "https://swagger.io/about"
            }
        },
        "OrderItem": {
            "required": [
                "order_id",
                "product_id",
                "file_id"
            ],
            "properties": {
                "id": {
                    "description": "Unique order item identifier",
                    "type": "string"
                },
                "order_id": {
                    "description": "Unique order Id related to existing order item",
                    "type": "string"
                },
                "product_id": {
                    "description": "Unique product Id related to existing order item",
                    "type": "string"
                },
                "file_id": {
                    "description": "Unique file Id related to existing order product_id",
                    "type": "string"
                }
            },
            "type": "object",
            "xml": {
                "name": "CartItem"
            },
            "externalDocs": {
                "description": "find more info here",
                "url": "https://swagger.io/about"
            }
        },
...
    },
...
}